Version 1 Puzzle List
I have included all the puzzles that will be in the Version 1 of our
game (Prototype). I will be implementing them over the course of
tomorrow and we should have the prototype ready to go by Friday /
Saturday. I have also updated the Rules Doc with the full list of
combinations for Addition/Subtraction.