Color change particle effect
Added the ability to display particles indicating color change. A
particle collider object must be created and then added to the color
gameplay script in order to activate it, no component, no effect, but no
crash. Also made having a text file to save the state of the scene being
played an optional feature.