Show hint tile
Hint tiles can now be displayed, hint tile object and color are given by
a variable and text file in color gameplay that follows the same format
as the victory grid. A button can be used to call the toggle draw hint
function and display the objects, they will not display by default.