Sound Manager
Included sound manager. To use, create the object and the buttons on the
starting screen. Then add the sound settings script to an audio source
in the scene. A button can call the draw sound options menu to draw the
manager at any time. Be sure to include a button that calls the draw
manager screen function from the sound manager script. Passing in a
false parameter will close the sound options display.