Bug fixes
Corrected the error where edges wouldn't render. Will need to adjust the
values of the variables Center to Bottom, Center to Left and Center to
Right variables in Create Grid in order to ensure they are drawn in the
right position . Values that seem best at the moment are 0.5 for bottom
and 0.25 for left and 0.25 right.